+++ src/felupe/thermal/__init__.py
@@ -1,11 +1,11 @@
+from ._solidbody_surface_heat_transfer import SolidBodySurfaceHeatTransfer
 from ._solidbody_thermal import SolidBodyThermal
-from ._solidbody_thermal_convection import SolidBodyThermalConvection
 from ._solidbody_thermal_heat_flux import SolidBodyThermalHeatFlux
 from ._time_step import TimeStep
 
 __all__ = [
     "SolidBodyThermal",
-    "SolidBodyThermalConvection",
+    "SolidBodySurfaceHeatTransfer",
     "SolidBodyThermalHeatFlux",
     "TimeStep",
 ]